src/requirements.txt
Flask
psycopg2-binary
python-dotenv
cryptography
requests
concurrent.futures